add cw-storage, refactor to multiple files
