+++ Cargo.toml
@@ -30,6 +30,9 @@ singlepass = [ "cosmwasm-vm/default-singlepass"]
 
 [dependencies]
 cosmwasm = { version = "~0.6.0" }
+cw-storage = "0.1.0"
+named_type = "=0.2.1"
+named_type_derive = "=0.2.1"
 schemars = "=0.5"
 serde = { version = "=1.0.103", default-features = false, features = ["derive"] }
 snafu = { version = "=0.5.0", default-features = false, features = ["rust_1_30"] }
