+++ examples/schema.rs
@@ -4,7 +4,8 @@ use std::path::PathBuf;
 
 use schemars::{schema::RootSchema, schema_for};
 
-use {{crate_name}}::contract::{CountResponse, HandleMsg, InitMsg, QueryMsg, State};
+use {{crate_name}}::msg::{CountResponse, HandleMsg, InitMsg, QueryMsg};
+use {{crate_name}}::state::State;
 
 fn main() {
     let mut pwd = current_dir().unwrap();
